requirements.txt
requests
python-dotenv
BeautifulSoup4
streamlit
replicate